Move gitweb module updates to repohook
- update naming from repo to repohook (see RR comments for reasoning)
- Commands output updated to escape markdown characters in routes and help
  (unescaped, these break Telegram support; initially added in 73fc6f61)
